requirements.txt
streamlit
PyMuPDF
sentence-transformers
faiss-cpu
google-generativeai
